fix: revert "feat: http upload/download progress handlers" (#58)
This reverts commit d30be969d84ad9428f344257be20e350db292b2f.
