+++ package.json
@@ -12,7 +12,6 @@
     "dist"
   ],
   "browser": {
-    "./src/http/fetch.js": "./src/http/fetch.browser.js",
     "./src/text-encoder.js": "./src/text-encoder.browser.js",
     "./src/text-decoder.js": "./src/text-decoder.browser.js",
     "./src/temp-dir.js": "./src/temp-dir.browser.js",
@@ -45,15 +44,15 @@
     "merge-options": "^2.0.0",
     "nanoid": "^3.1.3",
     "node-fetch": "^2.6.0",
-    "stream-to-it": "^0.2.0",
-    "it-to-stream": "^0.1.2"
+    "stream-to-it": "^0.2.0"
   },
   "devDependencies": {
     "aegir": "^25.0.0",
     "delay": "^4.3.0",
     "it-all": "^1.0.2",
     "it-drain": "^1.0.1",
-    "it-last": "^1.0.2"
+    "it-last": "^1.0.2",
+    "it-to-stream": "^0.1.2"
   },
   "contributors": [
     "Hugo Dias <hugomrdias@gmail.com>",
