--- src/http/error.js
@@ -1,26 +1,0 @@
-'use strict'
-
-class TimeoutError extends Error {
-  constructor (message = 'Request timed out') {
-    super(message)
-    this.name = 'TimeoutError'
-  }
-}
-exports.TimeoutError = TimeoutError
-
-class AbortError extends Error {
-  constructor (message = 'The operation was aborted.') {
-    super(message)
-    this.name = 'AbortError'
-  }
-}
-exports.AbortError = AbortError
-
-class HTTPError extends Error {
-  constructor (response) {
-    super(response.statusText)
-    this.name = 'HTTPError'
-    this.response = response
-  }
-}
-exports.HTTPError = HTTPError
